+++ components/dnd/card.tsx
@@ -1,11 +1,5 @@
-import React, { useRef } from "react";
-import {
-  useDrag,
-  useDrop,
-  DropTargetMonitor,
-  DragSourceMonitor,
-} from "react-dnd";
-import { XYCoord } from "dnd-core";
+import React from "react";
+import { useDraggableItem } from "../../hooks/useDraggableItem";
 
 export const ItemTypes = {
   CARD: "card",
@@ -21,81 +15,22 @@ const style = {
 };
 
 export interface CardProps {
-  id: any;
+  id: string;
   text: string;
   index: number;
   moveCard: (dragIndex: number, hoverIndex: number) => void;
 }
 
-interface DragItem {
-  index: number;
-  id: string;
-  type: string;
-}
-
 export const Card: React.FC<CardProps> = ({ id, text, index, moveCard }) => {
-  const ref = useRef<HTMLDivElement>(null);
-  const [, drop] = useDrop({
-    accept: ItemTypes.CARD,
-    hover(item: DragItem, monitor: DropTargetMonitor) {
-      if (!ref.current) {
-        return;
-      }
-      const dragIndex = item.index;
-      const hoverIndex = index;
-
-      // Don't replace items with themselves
-      if (dragIndex === hoverIndex) {
-        return;
-      }
-
-      // Determine rectangle on screen
-      const hoverBoundingRect = ref.current?.getBoundingClientRect();
-
-      // Get vertical middle
-      const hoverMiddleY =
-        (hoverBoundingRect.bottom - hoverBoundingRect.top) / 2;
-
-      // Determine mouse position
-      const clientOffset = monitor.getClientOffset();
-
-      // Get pixels to the top
-      const hoverClientY = (clientOffset as XYCoord).y - hoverBoundingRect.top;
-
-      // Only perform the move when the mouse has crossed half of the items height
-      // When dragging downwards, only move when the cursor is below 50%
-      // When dragging upwards, only move when the cursor is above 50%
-
-      // Dragging downwards
-      if (dragIndex < hoverIndex && hoverClientY < hoverMiddleY) {
-        return;
-      }
-
-      // Dragging upwards
-      if (dragIndex > hoverIndex && hoverClientY > hoverMiddleY) {
-        return;
-      }
-
-      // Time to actually perform the action
-      moveCard(dragIndex, hoverIndex);
-
-      // Note: we're mutating the monitor item here!
-      // Generally it's better to avoid mutations,
-      // but it's good here for the sake of performance
-      // to avoid expensive index searches.
-      item.index = hoverIndex;
-    },
-  });
-
-  const [{ isDragging }, drag] = useDrag({
-    item: { type: ItemTypes.CARD, id, index },
-    collect: (monitor: DragSourceMonitor) => ({
-      isDragging: monitor.isDragging(),
-    }),
-  });
+  const { ref, isDragging } = useDraggableItem(
+    index,
+    id,
+    ItemTypes.CARD,
+    moveCard
+  );
 
   const opacity = isDragging ? 0 : 1;
-  drag(drop(ref));
+
   return (
     <div ref={ref} style={{ ...style, opacity }}>
       {text}
