+++ pages/dnd.tsx
@@ -5,32 +5,32 @@ import { Container, Item } from "../components/dnd/container";
 
 const items: Item[] = [
   {
-    id: 1,
+    id: "id-1",
     text: "Write a cool JS library",
   },
   {
-    id: 2,
+    id: "id-2",
     text: "Make it generic enough",
   },
   {
-    id: 3,
+    id: "id-3",
     text: "Write README",
   },
   {
-    id: 4,
+    id: "id-4",
     text: "Create some examples",
   },
   {
-    id: 5,
+    id: "id-5",
     text:
       "Spam in Twitter and IRC to promote it (note that this element is taller than the others)",
   },
   {
-    id: 6,
+    id: "id-6",
     text: "???",
   },
   {
-    id: 7,
+    id: "id-7",
     text: "PROFIT",
   },
 ];
